Refactor testing from patch to pytest fixtures (#47)
* Refactor testing from mocks to pytest fixtures

* Update tests to use data directly instead of mocking

* Update CI for data fixture

* remove top level module export and refactor conftests (#48)

---------
