+++ space2stats_api/src/space2stats/__init__.py
@@ -1,5 +1,3 @@
 """space2stats."""
 
-from .main import fields, summaries  # noqa
-
 __version__ = "0.1.0"
